+++ lib/tankard/error.rb
@@ -3,6 +3,8 @@ module Tankard
     ConfigurationError = Class.new(::StandardError)
     NoBeerId = Class.new(::StandardError)
     NoStyleId = Class.new(::StandardError)
+    NoSearchQuery = Class.new(::StandardError)
+    MissingParameter = Class.new(::StandardError)
     HttpError = Class.new(::StandardError)
     LoadError = Class.new(::StandardError)
     InvalidResponse = Class.new(::StandardError)