+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Add params method to beer.  Allows passing additional parameters in an additional way.
 * Add styles method & class.  Allows querying of the /styles route
 * Add style method & class.  Allows querying of the /style/:id route
-# Add search method & class. Allows querying of the /search route
+* Add search method & class. Allows querying of the /search route
+* Add documentation for beer
 
 ## v0.0.1
 * Support for querying a beer